backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.8.0
GitPython>=3.1.43
tree-sitter>=0.23.2
tree-sitter-language-pack>=0.8.0
tree-sitter-typescript>=0.21.2
tree-sitter-javascript>=0.21.3
tree-sitter-python>=0.21.0
tree-sitter-go>=0.20.0
tree-sitter-java>=0.21.0
tree-sitter-c>=0.20.0
tree-sitter-cpp>=0.20.0
tree-sitter-c-sharp>=0.20.0
tree-sitter-ruby>=0.20.0
tree-sitter-php>=0.20.0
tree-sitter-rust>=0.21.0
tree-sitter-kotlin>=0.20.0
tree-sitter-swift==0.0.1
tree-sitter-html>=0.20.0
tree-sitter-css>=0.20.0
tree-sitter-json>=0.20.0
tree-sitter-yaml==0.7.2
tree-sitter-toml==0.7.0
tree-sitter-bash>=0.20.0
tree-sitter-lua==0.4.1
# tree-sitter-r has no Windows wheels; disabled for now
tree-sitter-sql==0.3.11
networkx>=3.3
faiss-cpu>=1.8.0
numpy>=2.0.0
requests>=2.32.0
PyJWT>=2.8.0
email-validator>=2.0.0
pytest>=8.2.0
keyring>=24.0.0
python-dotenv>=1.0.0
pydantic-ai>=1.0.6
